31 | Comment 1:N relationship

+++ articles/views.py
@@ -3,9 +3,12 @@ from django.shortcuts import render, redirect, get_object_or_404
 from django.contrib.auth.decorators import login_required
 from django.views.decorators.http import require_POST
 from django.contrib import messages
+# from django.core.exceptions import PermissionDenied
+from django.http import HttpResponseForbidden
 # 중간에 embed를 사용하면 ipython을 이용하여 중간 값을 확인할 수 있다.
 # from IPython import embed
 
+from django.contrib.auth import get_user_model
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
 
@@ -33,7 +36,9 @@ def create(request):
             # title = article_form.cleaned_data.get('title')
             # content = article_form.cleaned_data.get('content')
             # article = Article.objects.create(title=title, content=content)
-            article = article_form.save()
+            article = article_form.save(commit=False)
+            article.user = request.user
+            article.save()
             # context = {
             #     'article': article
             # }
@@ -68,12 +73,15 @@ def detail(request, article_pk):
 @require_POST
 # require_POST를 사용하면 if문을 사용하지 않아도 됌
 def delete(request, article_pk):
-    article = Article.objects.get(pk=article_pk)
-    # if request.method == 'POST':
-    article.delete()
-    return redirect('articles:index')
-    # else:
-    #     return redirect('articles:detail', article.pk)
+    if article.user == request.user:
+        article = Article.objects.get(pk=article_pk)
+        # if request.method == 'POST':
+        article.delete()
+        return redirect('articles:index')
+        # else:
+        #     return redirect('articles:detail', article.pk)
+    else:
+        raise PermissionDenied
 
 
 # def edit(request, article_pk):
@@ -85,22 +93,25 @@ def delete(request, article_pk):
 
 def update(request, article_pk):
     article = Article.objects.get(pk=article_pk)
-    if request.method == 'POST':
-        # instance값을 줘서 수정가능하도록
-        article_form = ArticleForm(request.POST, request.FILES, instance=article)
-        if article_form.is_valid():
-            # content = article_form.cleaned_data.get('content')
-            # article.content = content
-            # article.save()
-            article.image = request.FILES.get('image')
-            article = article_form.save()
-            return redirect('articles:detail', article.pk)
+    if article.user == request.user:
+        if request.method == 'POST':
+            # instance값을 줘서 수정가능하도록
+            article_form = ArticleForm(request.POST, request.FILES, instance=article)
+            if article_form.is_valid():
+                # content = article_form.cleaned_data.get('content')
+                # article.content = content
+                # article.save()
+                article.image = request.FILES.get('image')
+                article = article_form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            article_form = ArticleForm(instance=article)
+        context = {
+            'article_form': article_form
+        }
+        return render(request, 'articles/form.html', context)
     else:
-        article_form = ArticleForm(instance=article)
-    context = {
-        'article_form': article_form
-    }
-    return render(request, 'articles/form.html', context)
+        raise PermissionDenied
 
 @require_POST
 def comment_create(request, article_pk):
@@ -115,6 +126,7 @@ def comment_create(request, article_pk):
         comment = comment_form.save(commit=False)
         # 3-2. FK 넣고 저장
         comment.article = article
+        comment.user = request.user
         comment.save()
         messages.success(request, '댓글이 생성되었습니다.')
     else:
@@ -128,6 +140,10 @@ def comment_create(request, article_pk):
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
-    messages.success(request, '댓글이 삭제되었습니다.')
-    return redirect('articles:detail', article_pk)
+    if comment.user == request.user:
+        comment.delete()
+        messages.success(request, '댓글이 삭제되었습니다.')
+        return redirect('articles:detail', article_pk)
+    else:
+        # raise PermissionDenied 
+        return HttpResponseForbidden()
